moved Exceptions to error file
